Move mutation token types to their own directories
