+++ src/Hermes/Tokens/Mutations/Delete/class-delete-mutation-token.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Gravity_Forms\Gravity_Tools\Hermes\Tokens\Mutations;
+namespace Gravity_Forms\Gravity_Tools\Hermes\Tokens\Mutations\Delete;
 
 use Gravity_Forms\Gravity_Tools\Hermes\Tokens\Mutation_Token;
