+++ src/Hermes/Tokens/Mutations/Insert/class-insert-mutation-token.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Gravity_Forms\Gravity_Tools\Hermes\Tokens\Mutations;
+namespace Gravity_Forms\Gravity_Tools\Hermes\Tokens\Mutations\Insert;
 
 use Gravity_Forms\Gravity_Tools\Hermes\Tokens\Mutation_Token;
