+++ src/Hermes/Tokens/Mutations/Update/class-update-mutation-token.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Gravity_Forms\Gravity_Tools\Hermes\Tokens\Mutations;
+namespace Gravity_Forms\Gravity_Tools\Hermes\Tokens\Mutations\Update;
 
 use Gravity_Forms\Gravity_Tools\Hermes\Tokens\Mutation_Token;
